Silence two compiler warnings
git-svn-id: http://repo.ham.fi/svn/aprsc/trunk@664 3ce903b1-3385-4e86-93cd-f9a4a239f7ac

+++ src/historydb.c
@@ -115,7 +115,7 @@ void historydb_atend(void)
 
 void historydb_dump_entry(FILE *fp, struct history_cell_t *hp)
 {
-	fprintf(fp, "%ld\t", hp->arrivaltime);
+	fprintf(fp, "%ld\t", (long)hp->arrivaltime);
 	fwrite(hp->key, hp->keylen, 1, fp);
 	fprintf(fp, "\t");
 	fprintf(fp, "%d\t%d\t", hp->packettype, hp->flags);
